Merge pull request #17521 from bradcray/changes-contributors-1.24.1
CHANGES and CONTRIBUTORS files for 1.24.1

[reviewed by @ronawho , @e-kayrakli , @mppf, @dlongnecke-cray]

These are the CHANGES and CONTRIBUTORS files for 1.24.1. For future reference, the main motivators for the 1.24.1 release were the Infiniband performance improvements (which couldn't happen earlier due to the timeline in which we found and fixed them, combined with the GASNet release schedule) and the HPE Cray EX portability improvements. The other changes included here were selected due to being user-driven, low-risk, and not necessitating a re-spin of the documentation. We also rolled in the memory leaks fixes in order to achieve our first release with zero known memory leaks.