Revert "Move resolveSpecifiedReturnType()"
This reverts commit 2224ca95075a6d1b070b69cbed9a7e6467c83aba.

This commit caused an exceedingly large amount of errors of the form: "error:
'param' functions cannot return non-'param' values" for two tests of the Version
module, library/standard/Version/compareChplVersion.chpl and
library/standard/Version/compareChplVersionErrors.chpl when merged onto my
operators branch.  Since the original PR itself was uncertain of why the code
was in the previous state, it made sense to just undo the change.

Discussed with Vass.

Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>