Fix a case where AtomicObject would call memcpy with overflowing size
AtomicObject has a 'decompress' routine that unpacks a packed pointer and
returns it. It uses memcpy to copy an untyped wide pointer into a typed
one. This memcpy is only actually run if the locale field differs, which
should not happen in single locale compilation

At the same time, a C compiler doing static analysis of this memcpy would
notice that, in a single locale compilation, it overflows the destination
(since in that event the destination is only 8 bytes but it's copying
16).

This commit just includes '_local' in the earlier condition to make the
memcpy not be emitted for single-locale builds.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>