First cut at support for native AMOs in message-order-fence MCM mode.
This is the first cut at support for atomic operations in the new MCM
mode that uses message orderings plus fences.  It's similar to the
changes made to do the same thing for PUTs and GETs. The most apparent
difference is non-functional, and that is that in order to reduce the
argument list lengths for the necessarily non-inlined indirect calls,
instead of passing around addresses and MR keys and descriptors etc. we
bundle everything into a struct that can be used in a natural way when
calling fi_...atomicmsg(), and pass a pointer to that instead.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>