Remove support for CRC-checking received messages.
The comm layer's Active Message support should be stable enough at this
point that support for CRC-based integrity checks on received messages
is no longer useful.  I think it's been a year since I've used it, and
certainly nobody else has.  It can be brought back if needed; for now
remove it, because it is a little intrusive even though only available
under debug.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>