Add the cast operator to the list of operators checked in isAstrOpName
This allows the cast operator to be treated the same as the other operators when
it comes to determining whether calls to them should be transformed into method
calls when inside methods of types that happen to define their operator as a
method.

The map type was running into this, and also the set type.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>