Fix bug with isSubtype and equally generic child/parent types
Prior to this change, we would falsely report that a fully generic child type
was not a subtype of the fully generic parent type, but would report
appropriately that an instantiation of the child was a subtype of the parent.

This was because we weren't appropriately traversing when the child type was
generic.  To fix that, add a check before we look at if it is an instantiation
that will return the parent type, even if it is a direct match (rather than an
instantiation).

With this change, the test that failed before now passes (and it doesn't seem
to have broken the other ones I added)

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>