Store old results to prevent ABA-like problems
I have not observed such a problem with the combining but I think it is
possible. The map could contain a value that contains a pointer to
something old. If the map entry it depends upon it had no changes, it
could also have no changes. That could be a problem if the pointer value
was freed and then allocated in the time between checking the
depended-upon entry and the map entry itself.

Anyway keeping the old values around fits in with the
'context->collectGarbage()' mechanism. It should be possible to access
memory inside of an old value, but the old value will be changed by
combine calls, so if it's garbage the value won't match up anyway. As a
result, after the parseFile call, AST elements in values should only have
their pointers compared in combine calls.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>