Don't select msg-order-fence mode unless we can use bound tx contexts.
The new msg-order-fence MCM mode is of no use if we can't bind transmit
contexts, because lacking that we have to force memory visibility for
PUTs etc. right away, before we release the transmit context we used to
initiate them.  So here, adjust the set/check logic for msg-order-fence
mode so that along with the other requirements, a provider must support
bound contexts in order for us to pick it.

While here, also generally rework the bound-context logic, including
changing the variable naming and comments to refer to "bound" instead of
"fixed" transmit contexts.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>