At Lydia's suggestion, introduced a prediff to avoid ongoing maintenance here
I'm often lazy about writing prediff scripts because I'm not so good
at scripting, but since Lydia handed me one to use as a model, I
didn't have an excuse not to adopt it here.  I put it into the
util/test/ directory for other tests to link to and use symbolically
as well.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>