Update query framework to for changed dependencies
Consider the case where one query depends on another; say A depends on B.
Before this commit, when running A, we must choose to run A if B is out
of date. After this commit, we rerun B while choosing if we want to run
A, and as a result can consider whether or not the output of A actually
changed.

This required creating a virtual function to recompute a query result,
which in turn required tha the query result class store the arguments.
Since that would be redundant with the map keys, this commit changes
to storing the query results in a set, and making the result itself
a 'mutable' field (meaning it can be changed even if const). These
mutable fields must not form part of the "key" that the set uses for
hashing/comparison.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>