Fix issue where docs temporary directory wasn't getting cleaned up on error
The variable storing the temporary directory for chpldoc was local, which meant
that we'd only remove its contents after a successful run of the compiler.  This
change replaces that variable with a global variable stored in the same location
as the temporary directory used for normal operations, and adjusts the clean up
call at the exit of the compiler's operations to clean it up as well (which runs
regardless of if the compiler ran successfully or encountered an error).

With this change, I no longer see left over chpldoc temporary repos.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>