Stop rolling our own temporary dir and rely on mkdtemp to make it
At Elliot's suggestion, use mkdtemp to make a temporary directory, for both
chpldoc and chpl programs.  Because mkdtemp requires its strings end in six Xs,
alter the name we use from "tmpdirloc/chpl[doc]-username-pid.deleteme" to
"tmpdirloc/chpl[doc]-username.deleteme-XXXXXX" and remove all references to the
PID in this setup.  The username stuff seems reasonable to keep and the Xs will
be replaced with a unique enough set of characters for our purposes.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>