Use CHPL_LLVM=bundled for make install testing
Our testing for `make install` intentionally doesn't source our
common.bash scripts that set up paths to llvm and other things because
it's trying to test the install from a user perspective. We now have to
set CHPL_LLVM to some value, so choose bundled since that's what we're
expecting to become the default.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>