Fix some safeCasts that won't work with large uint indices
I'm thinking that we should probably do more of this offset arithmetic
with uints because either (a) its positive, which gives us more
breathing room or, (b) it's negative in which case the wraparounds
on ints aren't guaranteed anyway.  However, my attempt to do that
didn't land easily, and I think our current potential issues with
large arrays or wraparound are pre-existing, so I'm going to back
away from some of the previous changes I'd tried.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>