Extend _cond_test to support a method-based version
_cond_test — a routine used to determine whether or not an expression
can be part of a conditional test or not — has traditionally been a
standalone function.  However, this causes problems in import-based
cod where it can make the function not visible when an instance of its
type is.

The fix here is to support a method-based version of _cond_test which
will be visible given a value of the specified type regardless of
whether or not the containing module has been used or imported.
Rather than converting all existing _cond_test routines over to this
method, I preserved the standalone versions that are within auto-used
modules, and made the current error case attempt to call the method
if it exists to minimize the amount of code changing.

Ultimately, we should decide whether this is a capability that users
should be able to provide for their own types, and if so, how it
should _really_ be expressed.  This feels related to implicit
conversions, while also feeling somewhat operator-like (in that it's a
built-in capability; less so in terms of it not being visible to a
user).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>