Fix a size -> sizeAs case in ChapelArray; tighten up idxType check
This fixes a case in which 'size' was used in transferring arrays,
causing warnings when the idxType wasn't int by switching it to
sizeAs().

It also tightens up the check for whether or not the warning needs to
be printed to support the case when the routine is (now) used for
indices of arbitrary type, since that can come up in associative
domains.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>