Use the helper to check against operator names for visibility
We were determining if we needed to use method visibility rules based on a list
of known names that wouldn't normally register as methods (but were associated
with a type).  This list didn't include all the operators in it and was made
prior to the helper for checking all operator names.  Remove the operator names
from the list and add an additional check using the operator names helper where
we were traversing that list.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>