Fix default arguments for dynamically dispatched methods
This is an attempt to fix a problem reported in #7880 and #17649 in
which the default arguments for a dynamically dispatched routine use
the defaults for the static class type's version of the method rather
than the dynamic class's.  The issue was that the methods that the
compiler introduces to represent the default values were tagged as
being `param` which prevented them from being dynamically dispatched.

Big thanks to @mppf for providing the necessary context and pointers
to be able to solve this refreshingly quickly.

Resolves #7880
Resolves #17649

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>