Add variation of original test that works; and a simple case for completeness
Something that @cassella noted today that I hadn't realized before is
that the original default-argument-class-override.chpl test works if
the default argument is not a param, but is a var.  Here, I'm adding a
copy of that test, not because it's more likely to break than the
retired future here, but because it's interesting to me that it worked
previously (so if one but not the other were to break in the future,
that might be valuable to know).

I also added a third test that ensures that the call to the parent
type still works.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>