Change how chpl_comm_getenvMaxHeapSize() says heap size was not set.
Prior to this chpl_comm_getenvMaxHeapSize() returned size_t, which is an
unsigned type, with the special value 0 (zero) signifying that the heap
size was not specified in the environment.  In an upcoming change we are
going to default to creating a fixed heap in some cases and we'll want
to accept CHPL_RT_MAX_HEAP_SIZE=0 explicitly to indicate that we should
not do so.  Therefore here, change chpl_comm_getenvMaxHeapSize() to
return ssize_t instead, with a negative value signifying that no value
was set in the environment.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>