Fix bug with jumping to definition point of classes
Because class instances are wrapped in owned/shared/etc wrappers, we were
failing to resolve their operators - for other methods, we wouldn't find a
candidate so would fall back on the forwarding support, but for operators we
would find the default implementation and never get to looking at the forwarded
type.  This changes the implementation so that we look at the canonical type
in the first pass, allowing us to find those operator methods.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>