Fix c_ptr conditional checks for --baseline runs
My previous approach to supporting 'if c_ptr' in an import-based
setting relied on being able to call a method on a null c_ptr pointer
value, but this didn't work when we compile with --baseline.  Here,
I'm reverting the routine to a standalone function for this case, but
moving it from CPtr to ChapelBase, so that it's always available
(ChapelBase already used CPtr, so this doesn't add any new
inter-module dependencies).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>