Fix some issues in chpl__serialViewIter
This addresses a few issues in this code that Engin pointed out in
https://github.com/Cray/chapel-private/issues/1904:

* It gets away from calling dom.chpl_intToIdx() because we don't
  want to have to provide this as a method on all non-DR domains

* Similarly, it changes away from using dom.intIdxType for similar
  reasons.

Most uses of these routines in this file are fine (I believe) because
they're calling the local DR version of the routines.  However, this
is a routine used across a lot of distinct domain map types, so it's
important that it use the standalone versions of the routines rather
than relying on the convenience methods.

Filed a test locking in Engin's original test and a second version of
it that I noticed would go down a different stridable path, revealing
new errors.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>