Add another dynamic dispatch with default arguments problem case
While working on a simple demo for the bug fix in #17653, I realized
that it isn't actually fixed in as many cases as I'd thought.
Specifically, it seems that if there aren't "obvious" calls to the
dynamically dispatched routine, we still get the default value based
on the static type.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>