Default to a large fixed heap on systems we expect to benefit from it.
On the target platforms with the most capable networks (IB, e.g.) the
ofi comm layer cannot use the best-performing providers (verbs, e.g.)
unless there is a fixed heap.  But in the past the comm layer hasn't
created a fixed heap by default, meaning that users had to force it to
do so through the `CHPL_RT_MAX_HEAP_SIZE` environment variable.

Here, when a tentative provider search indicates that a fixed heap could
improve performance by enabling use of a better provider, create one by
default and make it large: 85% of the compute node's physical memory.
We have to use a tentative provider search because we usually need to
make this decision early, when initializing the memory layer, which is
done long before the bulk of comm layer initialization.  We need this so
early because some memory layers including the default jemalloc based
one initialize quite differently with and without a fixed heap.  And
"tentative" here means that the provider search uses just the base
libfabric capabilities that we require.  In particular we don't look at
anything related to MCM conformance like the full provider search that
comes later does.  Really, here we're just trying to predict whether or
not we'll end up selecting a provider that needs a fixed heap, and if so
make one by default.

Structurally, the major change here is that there is a new function to
set up the base (non-MCM related) provider hints, which is now called
both from provider selection proper and the new code that decides about
a fixed heap.  This also gets ride of the function that used to print
delayed heap-related messages; those are now done where that had been
called, in the function that does final provider-related tuning.

Relatedly, this adds a new debug keyword for heap-related output.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>