Add future about confusing error when using ops on uints of mixed size
This adds a future that demonstrates that we generate a confusing
error message when applying operations (subtraction here) to a pair of
'uint's of distinct size.  Specifically, it speaks in terms of not
being able to apply operators to a uint and a signed integer even
though there is no signed integer.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>