Add a new default value of 'unset' for CHPL_LLVM
If CHPL_LLVM is not set and no suitable 'system' or 'bundled' LLVM is found
then CHPL_LLVM will print as 'unset' from the chplenv scripts. If it is
'unset' the build will error out with the message asking the user to set it
to one of the three legal values.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>