Update test/deprecated/README, add a bunch of tests and another README
test/deprecated/README indicates that the files within it are meant to be
short-lived.  Update it so that the new directory,
test/deprecated/functionality, is excluded from that notion of temporariness.

Add a README to new directory test/deprecated/functionality indicating that it
is intended to test the deprecation of symbols rather than testing the
deprecation warning of *specific* symbols (which should live in the parent
directory.

Also add some tests locking in the expected behavior of:
- const
- modules
- param
- proc
- type
- var
declarations

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>