Add code to the parser and the AST to store deprecation messages and status
Adds a new flag, FLAG_DEPRECATED, and a deprecationMsg field to all Symbols.
Adjusts the parser so that symbol declarations go through a pathway that could
include the deprecation keyword and a potential message.  When the `deprecated`
keyword is found, add either the supplied message or a default one (if a message
was not supplied) and the flag to the symbol.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>