For comm=ofi with PMI2, have the launcher set number of KVS entries.
If we have to use the PMI key-value store for barrier, allgather, and
broadcast in the comm=ofi out-of-band code, we've seen many cases in
which its default size is insufficient for what we're doing with it.
Here set PMI_MAX_KVS_ENTRIES large enough for our needs on the launch
side, unless it's already set and larger, and then the real executable
on the compute node will inherit that.  I'm doing it this way because
I'm pretty sure this environment variable gets read in .init code in the
real executable, so setting it there will be too late.

As a necessary side effect, this adds a numLocales argument to the comm
layer implementation-specific chpl_comm_preLaunch() function called by
the launcher.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>