Fix field deprecation messages
If a field is deprecated, copy the deprecation message over to the accessor
functions we create for it.  Also, ensure that copies of functions also copy the
deprecation message over.

These changes, plus a modification to add the "In function 'main':" message,
are sufficient to allow the field test to pass.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>