Keep checks on for tests that are sensitive to range bounds checking
Here, I'm fixing my classic error of checking in tests that perform
runtime checks, that fail under --fast because the checks have been
turned off.  I took the approach of forcing checks on for each of
these cases.

The printModuleInitOrder and countDeadModules cases are a bit funny
because they are now sensitive to whether or not checking is on
because the only thing they require from Types.chpl is the safeCast()
check which is completely inlined when checks are off, but has a
helper routine that is not when checks are on.  So here, I just forced
the relevant checks to be on as well.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>