Use helper to get deprecation message so we can include the name
Instead of computing the default message in build, determine it when we try to
get the deprecation message for use - that occurs on the symbol itself, so we
have the name and can use it in the default message.  Replace all direct calls
to the field to instead call this helper function.

Also, in order to allow the enum constants test of the default message to work,
I had to fix the warning message to use the test's name (I had accidentally put
a stray 's' in the name that did not belong)

This allows a few more tests to pass

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>