Also output deprecation warning for intervening modules in import statement
When a module along the import path is deprecated, we should be sure to generate
its deprecation warning.

With this change (and a fix to the ordering of the warning messages),
handleModule.chpl now passes.  While here, add a test specifically of when the
next symbol in the import chain is a module (since handleModule was exercising
the last symbol being a symbol in a module rather than a module for this case)

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>