Deprecate the BigInteger enum `Round` in favor of a new enum `round`
This new enum also uses all lowercase for the constants instead of all caps.

This meant deprecating the functions that took `Round` as an argument in favor
of new versions that take `round`.  This impacted:
- two versions of div_q
- two versions of div_r
- two versions of div_qr
- div_q_2exp
- div_r_2exp

When deprecating, replace the default argument value with just a type
declaration to avoid potential conflicts when relying on the default value (such
calls will naturally switch to using the new type)

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>