Respect salloc/sbatch --cpus-per-task when launching with slurm-srun
When we compute the value for `--cpus-per-task` check to see if we're
within an existing salloc/sbatch allocation that has already set the
value and use that instead of trying to compute our own and likely try
to use use too many cores.

This value can still be lowered with `CHPL_LAUNCHER_CORES_PER_LOCALE` if
a user wants to use only part of their allocation for a chapel program
keeping the rest for something else.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>