While here, fix a couple of other places I noticed weren't adjusted
Checks for assigning nil to records were not adjusting for the possibility of
the assignment operator being an operator method (though I'm not entirely
convinced this case would trigger, as the assignment function I was looking at
didn't trigger the "is a record" portion because it's a ref argument even though
the type it is a ref to is a record).

Also, postFold double checking that a param is assigned multiple times was not
appropriately adjusting for operator method calls.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>