Only generate a deprecation warning if we're not already in a deprecated symbol
It can be annoying to get deprecated warnings in symbols that are themselves
deprecated - I suspect that in situations like this, part of the reason the
outer symbol is deprecated is because it used another deprecated symbol but had
to be kept around for users to transition.  In such cases, the deprecation
warning will be addressed by removing the outer deprecated symbol, so we don't
need to warn about it.

This impacted the BigInteger module when deprecating the Round enum and the
functions that took it as an argument.  Users of the BigInteger module would
get a warning for every reference to the Round enum in there, even though these
remaining references were to ease the transition to using the new version of
the deprecated functions.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>