Fix bug where we weren't suppressing warnings that are generated by arguments
When an argument in a deprecated function references a deprecated symbol, we
should still avoid generating the warning for the deprecated symbol.  However,
the check I made was too basic to catch that case.

I could potentially suppress the warning if anywhere in the declaration chain
was deprecated, but I think that will have a performance impact so would like
to avoid it if I can.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>