Fix an argument list error when registering memory.
I had coded the requestedKey as the 5th argument and the must-be-zero
offset as the 6th argument in the call to register memory, but actually
they were supposed to be the other way around.  Up until now these had
always both been zero anyway, so we didn't notice, but I'm working on
registering more regions and this bug popped out.  Fix it.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>