Fix a teensy bug in comm=ugni.
While adapting some code in the ugni comm layer to use in the ofi comm
layer, I spotted what I believe to be a bug.  It looks to me like the
code that collects read/write memory ranges out of the process memory
map could store the nul character that follows the path associated with
a range in the byte following the end of the pathname buffer.  Fix this.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>