Register static data segment, heap, and stack in addition to fixed heap.
In the ofi comm layer when we have a fixed heap, also register the
static data segment as well as the existing pages of the process's
original heap and stack.  This has a significant beneficial effect when
we communicate to or from these regions.  In a highly artificial setup
for an unrelated investigation on an HPE Cray EX system, with 4 Chapel
locales per compute node and a total of 64 locales (thus 16 nodes),
this reduced the time needed for module initialization from roughly 160s
to roughly 1s, much of this apparently because it allowed the use of
direct RMA instead of AM-mediated PUTs for broadcasting all the config
constants from locale 0 to the other locales.

Much of what's here was adapted from the same functionality in the ugni
comm layer.

While here I also did some cleanup in the memory table and registration
code, which was pretty old and in particular had not been exercised with
more than one registered region in a very long time.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>