Fix CHPL_LLVM=system for Mac Mojave users using llvm@11 from homebrew
When using llvm@11 from homebrew on Mac Mojave, and using the
version of clang/clang++ it provides with the system linker,
the error message:

```
ld: unknown option: -platform_version
```

was being produced, as has been noticed by several other projects
online.  A workaround for this is to tell clang to assume a specific
linker version via:

```
-mlinker-version=450
```

this PR does that by adding a check to `gather-clang-sysroot-arguments`
that checks the Mac OS version, and adds this option if it's 1.14.x,
permitting the build and compilations to pass.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>