Add explicit string casts for common/easy cases
When we do a cast-to-string (say, in an error message) and the type
doesn't support such a cast, we tend to call the general-purpose
stringify() routine which is very cool in that it converts any type to
a string, but is very expensive in that it does it by virtue of
writing the value to string memory using some pretty nontrivial code.

This PR takes a few common types that are compact and easy to write to
strings (ranges, rectangular domains, byteIndices, codepointIndices)
and injects string casts for them to avoid these code paths.  This
reverses the increase in generated code that stemmed from #17628 and
then some resulting in a ~4000 line savings in generated code for
the stream benchmark with checks on.

I believe this will https://github.com/Cray/chapel-private/issues/2074
but am not marking it as such here until we verify in nightly testing.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>