Improve warning when deprecated config set at compilation time
In the execution flag case, we automatically indicate that a flag was the
problem and which flag.  But in the compilation flag case, we've lost
information about which flag (I believe in part because it's intermingled with
compilation specific flags, but also definitely because the deprecation warning
happens after the config is parsed rather than at the time).  This meant that
the warning generated wasn't as helpful, so add a note after the warning
indicating that the deprecated config was set via a flag.

Update the test that tracks this to stay in step with the change

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>