Fix bug with finding tertiary operator methods on classes
Turns out we were also failing to check the canonical type's name when filtering
use/import statements.  Because we normally find operator candidates in our
initial round of checks, this meant that we would fail to resolve tertiary
operators on classes because we weren't checking the forwarded type name ever.
Resolve that issue (but only for operator calls, where it is most likely to be
relevant)

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>