Squash all deprecation warnings in deprecated symbols, no matter how deeply
We were squashing deprecation warnings that were defined directly in a
deprecated symbol, but would only check the immediate parent symbol.  This meant
that if a deprecated symbol had a use of another deprecated symbol in a much
deeper context, we'd still get those warnings.  This change makes it so that all
deprecated warnings in already deprecated symbols are squashed, no matter how
deeply nested.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>