Store chpl::ID in astlocT
We could find the filename and line number for uAST elements when
constructing the (old) AST for them. However the uAST is designed to
separate the filename and line number information from the other contents
of the source code in order to allow more incremental reuse. So it seems
unfortunate to compute the line numbers for everything when converting to
old AST.

Instead, this commit adjusts astlocT to store a chpl::ID that can be used
to retrieve the filename and line number when they are needed. Retrieving
these will set the filename and line number fields in the class (in case
caching them is useful). This allows us to put off finding the line
number and file name until there is an actual error. That should allow
more incremental reuse (once we get to reusing the old AST in some form)
and reduce the amount of lookups in the common case of correct code.

However there are still many uses of the location (especially getting the
file name) in the compiler sources that are not connected to reporting
errors. Hopefully many of these will go away as we migrate more of the
front end of the compiler to the uAST.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>