Stop the launcher from generating deprecation warnings about execution configs
We were accidentally generating a warning message for deprecated symbols in both
the launcher and the executable, leading to duplicate deprecation messages when
running with CHPL_LAUNCHER != none.  This change threads knowledge of whether
we are in the launcher through to the code that checks on configs, so that we
no longer warn from the launcher.

However, we will still generate a warning message per locale instead of per
program.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>