Don't try to use CQ wait/poll on macOS in comm=ofi.
The libfabric wait/poll support requires `epoll`, which is a Linux thing
rather than a POSIX one and thus isn't present on macOS.  But because
macOS is a rather secondary target for libfabric, they haven't coded any
sort of workaround for that either.  The result is that depending on the
circumstances you can see a variety of unpleasant behaviors (internal
error, hang, segfault) if you turn on waiting and/or polling for CQs. So
here, just avoid wait/poll with CQs entirely under macOS.

There is some background information in the comments on this libfabric
issue, though the issue itself isn't directly about this problem:
  https://github.com/ofiwg/libfabric/issues/5453

While here, I also fixed an incorrect format specifier in one of the
debug output fprintf() calls.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>