Make the saveSphinxInDocs test robust to changes to the Sphinx version
The .bad for this future contains the Sphinx version number as part of the
output, but we may update our version of Sphinx again before we resolve this
future.  Add a .prediff to filter out the version number from that output.

Update the .future file to indicate the prediff should probably be removed when
the future is resolved

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>